Update docs and use less memory for gradle tests
