+++ sqldelight-gradle-plugin/src/test/kotlin/com/squareup/sqldelight/properties/MultiModuleTests.kt
@@ -1,4 +1,4 @@
-package com.squareup.sqldelight.tests
+package com.squareup.sqldelight.properties
 
 import com.google.common.truth.Truth.assertThat
 import com.squareup.sqldelight.androidHome
